### 2025-07-07 e7a6b37
Fixed bibtex issue
