+++ paper.bib
@@ -1,9 +1,9 @@
 @article{Cauchy:1847,
   	Author = {{Cauchy}, A.},
-  	Journal = {Compte rendu des séances de l'académie des sciences},
+  	Journal = {{Compte rendu des séances de l'académie des sciences}},
   	Title = {{Méthode générale pour la résolution de systèmes d'équations simultanées}},
-  	Year = 1847,
-    pages = 536-538,
+  	Year = {1847},
+    pages = {536 -- 538},
 }
 
 @Book{Goodfellow:2016,
