### 2025-07-07 f96bf85
Update LICENSE.md to MIT license
