### 2025-11-17 #288 @ahudde
@openjournals/jose-editors Am I missing some step before an Editor is assigned to this submission?
