### 2025-11-17 b972c01
Corrected typo in bibliography
