+++ paper.bib
@@ -3,7 +3,7 @@
   	Journal = {{Compte rendu des séances de l'académie des sciences}},
   	Title = {{Méthode générale pour la résolution de systèmes d'équations simultanées}},
   	Year = {1847},
-    pages = {536 -- 538},
+    pages = {536--538},
 }
 
 @Book{Goodfellow:2016,
