### 2026-01-20 #288 @victoraalves
Thank you @fvlima-codes. @ashleefv, if you still need reviewers, I am more than happy to review this one. I'm also a reviewer for JOSS.
