### 2026-02-08 #323 @editorialbot
[REVIEW]: An Interactive Notebook for Learning Nonlinear 1 Programming with Constraints
**Submitting author:** @ahudde (<a href="http://orcid.org/0000-0002-5652-2815">Anselm Hudde</a>)
**Repository:** https://github.com/ahudde/nonlinear_programming
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
