### 2026-02-27 63ee590
Add status JOSE badge to README
