### 2026-04-21 #323 @ahudde
Dear @jindavid, dear @victoraalves,

I would like to ask, how the revision process is proceeding. Are there currently any tasks for me that I might have missed?

Yours,
Anselm
