### 2026-04-21 #323 @victoraalves
@ahudde Thank you for reaching out. In my opinion, the proposed Jupyter Notebook needs a major overhaul before being considered for publication at JOSE. Here are some points that might help enhancing your material:

Overall comment:
The learning module is of an overall nice quality, and contains a good set of analyzes and visualizations, covering most of the points of the review checklist above. However, it feels a bit short. Given how topic-rich is the constrained optimization field is, the learning module could explore more topics such as KKT conditions, SQP and interior point methods, for instance. I understand that this is ultimately a decision that depends on the context of the module within the scope of the class, but nevertheless the learning could be more substantial.

Regarding `nonlinear_programming.ipynb`:
1. Learning Objectives: Please have the Learning Objectives that you'd like to address in this tutorial/session. It is also interesting to have, at the end, a quick recap of what was covered.
2. The "Example: Increase in efficiency through Automatic Differentiation" section has no comments at the end or any discussion. Just code at the end. In my opinion, it would be interesting to have a more in-depth analysis/discussion of the final results.

Regarding the manuscript:

1. You have a missing reference on the "Statement of Need" section:

Please double check all references, including if they have DOIs associated with them. This is a must.

2. "Description, Instructional Design, and Project Story": Although this section contains a good description and instructional design, it could benefit from more details on the project story.

Regarding the repository:

1. Maybe having "Releases" so you can have versions on the repository.

I hope this can help going forward. My current assessment is Major Review, @ashleefv.
