### 2026-05-04 82648a4
@victoraalves: Fixed technical Issue with citation of Hastie 2009
