+++ paper.bib
@@ -21,7 +21,7 @@
   Author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
   editor = {},
   publisher = {Springer},
-  year = {2008}
+  year = {2009}
 }
 
 @book{Suykens:2014,
