### 2026-05-04 2bb2577
Cleaned up bibliography.bib
