+++ paper.bib
@@ -13,13 +13,12 @@
   Year                     = {2016},
 
   Address                  = {Cambridge, MA, USA},
-  Note                     = {\url{http://www.deeplearningbook.org}}
+  url = {http://www.deeplearningbook.org},
 }
 
 @book{Hastie:2009,
   Title = {The Elements of Statistical Learning},
   Author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
-  editor = {},
   publisher = {Springer},
   year = {2009}
 }
@@ -27,17 +26,17 @@
 @book{Suykens:2014,
   Title = {Regularization, Optimization, Kernels, and Support Vector Machines (1st ed.)},
   Author = {Suykens, J.A.K. and Signoretto, M. and Argyriou, A.},
-  editor = {},
   publisher = {Chapman and Hall/CRC},
   year = {2014}
 }
 
 @inproceedings{Daskalakis:2018,
-  title={The Limit Points of (Optimistic) Gradient Descent in Min-Max Optimization},
-  author={Constantinos Daskalakis and Ioannis Panageas},
-  booktitle={Neural Information Processing Systems},
-  year={2018},
-  url={https://api.semanticscholar.org/CorpusID:54170344}
+ author = {Daskalakis, Constantinos and Panageas, Ioannis},
+ booktitle = {Advances in Neural Information Processing Systems},
+ title = {The Limit Points of (Optimistic) Gradient Descent in Min-Max Optimization},
+ url = {https://proceedings.neurips.cc/paper_files/paper/2018/file/139c3c1b7ca46a9d4fd6d163d98af635-Paper.pdf},
+ volume = {31},
+ year = {2018}
 }
 
 @book{Taha:2017,
@@ -46,7 +45,6 @@
   publisher = {Pearson},
   year     = {2017},
   isbn     = {9781292165547},
-  doi      = {},
   url      = {https://elibrary.pearson.de/book/99.150005/9781292165561}
 }
 
