### 2026-05-05 72e2dfc
minor improvements in the code
