### 2026-05-05 e5ed31e
minor edits to improve readability of code
