### 2026-05-05 f9849da
minor code improvement
