### 2026-05-05 54c419c
Improved code
