+++ paper.bib
@@ -1,56 +1,60 @@
 @article{Cauchy:1847,
-  	Author = {{Cauchy}, A.},
-  	Journal = {{Compte rendu des séances de l'académie des sciences}},
-  	Title = {{Méthode générale pour la résolution de systèmes d'équations simultanées}},
-  	Year = {1847},
-    pages = {536--538},
+Author = {{Cauchy}, A.},
+Journal = {{Compte rendu des séances de l'académie des sciences}},
+Title = {{Méthode générale pour la résolution de systèmes d'équations simultanées}},
+Year = {1847},
+pages = {536--538},
 }
 
 @Book{Goodfellow:2016,
-  Title                    = {Deep Learning},
-  Author                   = {Ian J. Goodfellow and Yoshua Bengio and Aaron Courville},
-  Publisher                = {MIT Press},
-  Year                     = {2016},
-
-  Address                  = {Cambridge, MA, USA},
-  url = {http://www.deeplearningbook.org},
+Title = {Deep Learning},
+Author = {Ian J. Goodfellow and Yoshua Bengio and Aaron Courville},
+Publisher = {MIT Press},
+Year = {2016},
+Address = {Cambridge, MA, USA},
+url = {http://www.deeplearningbook.org},
+doi = {10.1007/s10710-017-9314-z},
 }
 
 @book{Hastie:2009,
-  Title = {The Elements of Statistical Learning},
-  Author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
-  publisher = {Springer},
-  year = {2009}
+Title = {The Elements of Statistical Learning},
+Author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
+publisher = {Springer},
+year = {2009},
+doi = {10.1007/978-0-387-84858-7};
 }
 
 @book{Suykens:2014,
-  Title = {Regularization, Optimization, Kernels, and Support Vector Machines (1st ed.)},
-  Author = {Suykens, J.A.K. and Signoretto, M. and Argyriou, A.},
-  publisher = {Chapman and Hall/CRC},
-  year = {2014}
+Title = {Regularization, Optimization, Kernels, and Support Vector Machines (1st ed.)},
+Author = {Suykens, J.A.K. and Signoretto, M. and Argyriou, A.},
+publisher = {Chapman and Hall/CRC},
+year = {2014},
+doi = {10.1201/b17558},
 }
 
 @inproceedings{Daskalakis:2018,
- author = {Daskalakis, Constantinos and Panageas, Ioannis},
- booktitle = {Advances in Neural Information Processing Systems},
- title = {The Limit Points of (Optimistic) Gradient Descent in Min-Max Optimization},
- url = {https://proceedings.neurips.cc/paper_files/paper/2018/file/139c3c1b7ca46a9d4fd6d163d98af635-Paper.pdf},
- volume = {31},
- year = {2018}
+author = {Daskalakis, Constantinos and Panageas, Ioannis},
+booktitle = {Advances in Neural Information Processing Systems},
+title = {The Limit Points of (Optimistic) Gradient Descent in Min-Max Optimization},
+url = {https://proceedings.neurips.cc/paper_files/paper/2018/file/139c3c1b7ca46a9d4fd6d163d98af635-Paper.pdf},
+volume = {31},
+year = {2018},
+doi = {0.5555/3327546.3327597},
 }
 
 @book{Taha:2017,
-  author   = {Taha, Hamdy},
-  title    = {{Operations Research An Introduction, Global Edition}},
-  publisher = {Pearson},
-  year     = {2017},
-  isbn     = {9781292165547},
-  url      = {https://elibrary.pearson.de/book/99.150005/9781292165561}
[…]
