### 2026-05-19 0b43d86
Corrected typo in paper.bib
