+++ paper.bib
@@ -21,7 +21,7 @@ Title = {The Elements of Statistical Learning},
 Author = {Trevor Hastie and Robert Tibshirani and Jerome Friedman},
 publisher = {Springer},
 year = {2009},
-doi = {10.1007/978-0-387-84858-7};
+doi = {10.1007/978-0-387-84858-7},
 }
 
 @book{Suykens:2014,
