+++ paper.bib
@@ -39,7 +39,6 @@ title = {The Limit Points of (Optimistic) Gradient Descent in Min-Max Optimizati
 url = {https://proceedings.neurips.cc/paper_files/paper/2018/file/139c3c1b7ca46a9d4fd6d163d98af635-Paper.pdf},
 volume = {31},
 year = {2018},
-doi = {0.5555/3327546.3327597},
 }
 
 @book{Taha:2017,
