### 2026-05-20 93074c0
Extended LICENSE to inlude text and figures
