+++ paper.bib
@@ -13,7 +13,7 @@ Publisher = {MIT Press},
 Year = {2016},
 Address = {Cambridge, MA, USA},
 url = {http://www.deeplearningbook.org},
-doi = {10.1007/s10710-017-9314-z},
+isbn = {978-0-262-03561-3},
 }
 
 @book{Hastie:2009,
@@ -46,7 +46,7 @@ author = {Taha, Hamdy},
 title = {{Operations Research An Introduction, Global Edition}},
 publisher = {Pearson},
 year = {2017},
-isbn = {9781292165547},
+isbn = {978-1-292-16556-1},
 url = {https://elibrary.pearson.de/book/99.150005/9781292165561}
 }
 
