### 2026-07-24 d4ebc74
Fix citation formatting in Statement of Need
