### 2026-08-18 #323 @ashleefv
> Hi [@ahudde](https://github.com/ahudde)! Yes, all my feedback has been addressed. I now recommend this learning module for publication in JOSE.

@jindavid to document this completely, can you fill in any remaining items from your reviewer checklist in this thread?
